test: add the integration test for `examples/fibonacci`, remove `fib`
test in `apps` integration tests
