+++ tests/integration/expected/examples/fib.wat
@@ -1,6 +1,6 @@
-(module $miden_integration_tests_rust_fib_wasm.wasm
+(module $fibonacci.wasm
   (type (;0;) (func (param i32) (result i32)))
-  (func $fib (;0;) (type 0) (param i32) (result i32)
+  (func $entrypoint (;0;) (type 0) (param i32) (result i32)
     (local i32 i32 i32)
     i32.const 0
     local.set 1
@@ -30,10 +30,6 @@
   )
   (memory (;0;) 16)
   (global $__stack_pointer (;0;) (mut i32) i32.const 1048576)
-  (global (;1;) i32 i32.const 1048576)
-  (global (;2;) i32 i32.const 1048576)
   (export "memory" (memory 0))
-  (export "fib" (func $fib))
-  (export "__data_end" (global 1))
-  (export "__heap_base" (global 2))
+  (export "entrypoint" (func $entrypoint))
 )
