--- tests/integration/expected/fib.hir
@@ -1,39 +1,0 @@
-(component 
-    ;; Modules
-    (module #miden_integration_tests_rust_fib_wasm
-        ;; Constants
-        (const (id 0) 0x00100000)
-
-        ;; Global Variables
-        (global (export #__stack_pointer) (id 0) (type i32) (const 0))
-        (global (export #gv1) (id 1) (type i32) (const 0))
-        (global (export #gv2) (id 2) (type i32) (const 0))
-
-        ;; Functions
-        (func (export #fib) (param i32) (result i32)
-            (block 0 (param v0 i32)
-                (let (v2 i32) (const.i32 0))
-                (let (v3 i32) (const.i32 0))
-                (let (v4 i32) (const.i32 1))
-                (br (block 2 v4 v0 v3)))
-
-            (block 1 (param v1 i32))
-
-            (block 2 (param v6 i32) (param v7 i32) (param v9 i32)
-                (let (v8 i1) (neq v7 0))
-                (condbr v8 (block 4) (block 5)))
-
-            (block 3 (param v5 i32))
-
-            (block 4
-                (let (v10 i32) (const.i32 -1))
-                (let (v11 i32) (add.wrapping v7 v10))
-                (let (v12 i32) (add.wrapping v9 v6))
-                (br (block 2 v12 v11 v6)))
-
-            (block 5
-                (ret v9))
-        )
-    )
-
-)
