--- tests/rust-apps-wasm/fib/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "miden-integration-tests-rust-fib-wasm"
-version = "0.0.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-dlmalloc = { version = "0.2.4", features = ["global"] }
-miden-integration-tests-rust-fib = { path = "../../rust-apps/fib" }
-
-[profile.release]
-opt-level = "z"
-debug = true
