--- tests/rust-apps/fib/src/lib.rs
@@ -1,13 +1,0 @@
-#![no_std]
-
-#[no_mangle]
-pub fn fib(n: u32) -> u32 {
-    let mut a = 0;
-    let mut b = 1;
-    for _ in 0..n {
-        let c = a + b;
-        a = b;
-        b = c;
-    }
-    a
-}
